tested superficially

+++ ft_putstr_fd.c
@@ -6,13 +6,13 @@
 /*   By: openculturalcenter <openculturalcenter@    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/17 12:56:58 by ifridrik          #+#    #+#             */
-/*   Updated: 2023/03/24 15:04:40 by opencultura      ###   ########.fr       */
+/*   Updated: 2023/03/24 15:24:37 by opencultura      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void ft_putstr_fd(char *s, int fd)
+void	ft_putstr_fd(char *s, int fd)
 {
-	write(fd, s, ft_strlen(s));
+	write(fd, s, (int) ft_strlen(s));
 }
